requirements.txt
apache-airflow==2.8.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2
python-dotenv==1.0.0
pyyaml==6.0.1
google-cloud-storage==2.10.0
dropbox==11.36.2
boto3==1.34.0
matplotlib==3.7.2
seaborn==0.12.2
